+++ es5/utils/parse.js
@@ -381,6 +381,41 @@ function getTargetSelector(target, context) {
   return target;
 }
 
+/**
+ * Parse the target event names
+ *
+ * @param {Array|String} eventNames e.g. `Component:customEvent dom:mouseover`
+ * @param {String} namespace Optional namespace to assign each extracted custom (non-DOM) event name
+ * @returns {Array}
+ */
+function extractTargetEventNames(inputEventNames, namespace) {
+  var targetEventNames = [];
+  var eventNames = inputEventNames;
+
+  if (typeof inputEventNames === 'undefined' ? 'undefined' : _typeof(inputEventNames)) {
+    // Split eventNames by spaces
+    if (/\s/.test(inputEventNames)) {
+      eventNames = inputEventNames.split(/\s+/);
+    }
+  }
+
+  // Process each event name
+  eventNames.forEach(function (eventName) {
+    // Default to namespaced event name
+    var targetEventName = typeof namespace === 'string' && namespace !== '' ? namespace + ':' + eventName : eventName;
+
+    // Remove any reference to the native DOM event namespace
+    if (/^dom:/i.test(eventName)) {
+      targetEventName = eventName.replace(/^dom\:/gi, '', eventName);
+    }
+
+    // Add to the list
+    targetEventNames.push(targetEventName);
+  });
+
+  return targetEventNames;
+}
+
 var parse = {
   coerceToPrimitiveType: coerceToPrimitiveType,
   convertToBoolean: convertToBoolean,
@@ -390,7 +425,8 @@ var parse = {
   extractTriggerDetails: extractTriggerDetails,
   fixedEncodeURIComponent: fixedEncodeURIComponent,
   getTargetBySelector: getTargetBySelector,
-  getTargetSelector: getTargetSelector
+  getTargetSelector: getTargetSelector,
+  extractTargetEventNames: extractTargetEventNames
 };
 
 module.exports = parse;